refactor(storage): move memory and file to local
